tools/requirements.txt
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.1
fake-useragent==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pyproj==3.6.1
python-magic==0.4.27
sniffio==1.3.1
typing_extensions==4.11.0
